Add regularized lr impl and related unit tests
