+++ tests/test_lin_rg.py
@@ -1,0 +1,450 @@
+import unittest
+import os
+
+
+from numpy.testing import assert_allclose
+from numpy import ones, zeros, int64, array, append, genfromtxt
+
+from ml_algorithms.lin_rg import (normal_eqn, cost_func,
+                                  cost_func_reg, grad,
+                                  reg_grad, predict, h)
+
+TESTDATA1 = os.path.join(os.path.dirname(__file__), 'data1.csv')
+TESTDATA2 = os.path.join(os.path.dirname(__file__), 'data2.csv')
+
+
+class TestLinearRegression(unittest.TestCase):
+
+    def setUp(self):
+        self.data1 = genfromtxt(TESTDATA1, delimiter=',')
+        self.data2 = genfromtxt(TESTDATA2, delimiter=',')
+
+# NORMAL EQUATION
+
+    def test_normal_eqn_data1(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int)
+        X = append(intercept, X, axis=1)
+
+        assert_allclose([[-3.896], [1.193]],
+                        normal_eqn(X, y),
+                        rtol=0, atol=0.001)
+
+    def test_normal_eqn_data2(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int)
+        X = append(intercept, X, axis=1)
+
+        assert_allclose([[89597.909], [139.210], [-8738.019]],
+                        normal_eqn(X, y),
+                        rtol=0, atol=0.001)
+
+# COST FUNCTION
+
+    def test_cost_func_data1_1(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = zeros((n + 1, 1), dtype=int64)
+
+        assert_allclose([[32.073]],
+                        cost_func(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_cost_func_data1_2(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+
+        assert_allclose([[10.266]],
+                        cost_func(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_cost_func_data1_3(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-1], [2]])
+
+        assert_allclose([[54.242]],
+                        cost_func(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_cost_func_data2_1(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = zeros((n + 1, 1), dtype=int64)
+
+        assert_allclose([[65591548106.457]],
+                        cost_func(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_cost_func_data2_2(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+
+        assert_allclose([[64828197300.798]],
+                        cost_func(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_cost_func_data2_3(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-25.3], [32], [7.8]])
+
+        assert_allclose([[43502644952.311]],
+                        cost_func(X, y, theta),
+                        rtol=0, atol=0.001)
+
+# REGULARIZED COST FUNCTION
+
+    def test_reg_cost_func_data1_1(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+        _lambda = 0
+
+        assert_allclose([[10.266]],
+                        cost_func_reg(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_reg_cost_func_data1_2(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+        _lambda = 100
+
+        assert_allclose([[10.781984]],
+                        cost_func_reg(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_reg_cost_func_data1_3(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-1], [2]])
+        _lambda = 750
+
+        assert_allclose([[69.706373]],
+                        cost_func_reg(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_cost_func_reg_data2_1(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+        _lambda = 0
+
+        assert_allclose([[64828197300.798]],
+                        cost_func_reg(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_cost_func_reg_data2_2(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+        _lambda = 1000000
+
+        assert_allclose([[64828218577.393623]],
+                        cost_func_reg(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_cost_func_reg_data2_3(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-25.3], [32], [7.8]])
+        _lambda = 1000000
+
+        assert_allclose([[43514185803.375198]],
+                        cost_func_reg(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+# GRADIENT
+
+    def test_grad_data1_1(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = zeros((n + 1, 1), dtype=int64)
+
+        assert_allclose([[-5.839], [-65.329]],
+                        grad(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_grad_data1_2(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+
+        assert_allclose([[3.321], [24.235]],
+                        grad(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_grad_data1_3(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-1], [2]])
+
+        assert_allclose([[9.480], [89.319]],
+                        grad(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_grad_data2_1(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = zeros((n + 1, 1), dtype=int64)
+
+        assert_allclose([[-340412.659], [-764209128.191], [-1120367.702]],
+                        grad(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_grad_data2_2(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+
+        assert_allclose([[-338407.808], [-759579615.064], [-1113679.894]],
+                        grad(X, y, theta),
+                        rtol=0, atol=0.001)
+
+    def test_grad_data2_3(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-25.3], [32], [7.8]])
+
+        assert_allclose([[-276391.445], [-616340858.434], [-906796.414]],
+                        grad(X, y, theta),
+                        rtol=0, atol=0.001)
+
+# REGULARIZED GRADIENT
+
+    def test_reg_grad_data1_1(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+        _lambda = 0
+
+        assert_allclose([[3.321], [24.235]],
+                        reg_grad(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_reg_grad_data1_2(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+        _lambda = 100
+
+        assert_allclose([[3.320665], [25.265821]],
+                        reg_grad(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_reg_grad_data1_3(self):
+        y = self.data1[:, -1:]
+        X = self.data1[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-1], [2]])
+        _lambda = 750
+
+        assert_allclose([[9.480465], [104.783153]],
+                        reg_grad(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_reg_grad_data2_1(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+        _lambda = 0
+
+        assert_allclose([[-338407.808], [-759579615.064], [-1113679.894]],
+                        reg_grad(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_reg_grad_data2_2(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+        _lambda = 1000000
+
+        assert_allclose([[-338407.808], [-759558338.468], [-1092403.298]],
+                        reg_grad(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+    def test_reg_grad_data2_3(self):
+        y = self.data2[:, -1:]
+        X = self.data2[:, :-1]
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-25.3], [32], [7.8]])
+        _lambda = 1000000
+
+        assert_allclose([[-276391.444681],
+                         [-615660007.370213],
+                         [-740838.968085]],
+                        reg_grad(X, y, theta, _lambda),
+                        rtol=0, atol=0.001)
+
+# PREDICT
+
+    def test_predict_1(self):
+        X = array([[3.5]])
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-3.6303], [1.1664]])
+
+        assert_allclose([[0.4521]],
+                        predict(X, theta),
+                        rtol=0, atol=0.001)
+
+    def test_predict_2(self):
+        X = array([[3.5]])
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = zeros((n + 1, 1), dtype=int64)
+
+        assert_allclose([[0]],
+                        predict(X, theta),
+                        rtol=0, atol=0.001)
+
+    def test_predict_3(self):
+        X = array([[-3.5, 2.7]])
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+
+        assert_allclose([[0.2]],
+                        predict(X, theta),
+                        rtol=0, atol=0.001)
+
+    def test_predict_4(self):
+        X = array([[-3.5, 2.7]])
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = -1 * ones((n + 1, 1), dtype=int64)
+
+        assert_allclose([[-0.2]],
+                        predict(X, theta),
+                        rtol=0, atol=0.001)
+
+# HYPOTHESYS
+
+    def test_h_1(self):
+        X = array([[3.5]])
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = array([[-3.6303], [1.1664]])
+
+        assert_allclose([[0.4521]],
+                        h(X, theta),
+                        rtol=0, atol=0.001)
+
+    def test_h_2(self):
+        X = array([[3.5]])
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = zeros((n + 1, 1), dtype=int64)
+
+        assert_allclose([[0]],
+                        h(X, theta),
+                        rtol=0, atol=0.001)
+
+    def test_h_3(self):
+        X = array([[-3.5, 2.7]])
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = ones((n + 1, 1), dtype=int64)
+
+        assert_allclose([[0.2]],
+                        h(X, theta),
+                        rtol=0, atol=0.001)
+
+    def test_h_4(self):
+        X = array([[-3.5, 2.7]])
+        m, n = X.shape
+        intercept = ones((m, 1), dtype=int64)
+        X = append(intercept, X, axis=1)
+        theta = -1 * ones((n + 1, 1), dtype=int64)
+
+        assert_allclose([[-0.2]],
+                        h(X, theta),
+                        rtol=0, atol=0.001)
+
+
+if __name__ == '__main__':
+    unittest.main()
